Refactor scanner backend and update documentation
- Updated `Readme.md` with scanner setup instructions.
- Removed x64 configurations from `TownSuite.TwainScanner.sln`.
- Introduced new class structure in `IBackend.cs` for scanning and saving images in multiple formats with OCR support.
- Enhanced `Naps2TwainBackend.cs` and `OriginalTwainBackend.cs` for improved scanning functionality.
- Modified `OriginalWiaBackend.cs` for better WIA scanning management.
- Significant UI updates in `MainFrame.Designer.cs` to accommodate new features.
- Refactored `MainFrame.cs` to support dynamic backend selection (TWAIN, WIA, NAPS2).
- Made `Enabled` property public in `Ocr.cs` for better OCR management.
- Updated `Program.cs` to accept command-line arguments for backend selection.
- Changed `TwainScanner.csproj` to target .NET 8.0 and added necessary package references.
