+++ TownSuite.TwainScanner.sln
@@ -7,18 +7,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TwainScanner", "TownSuite.T
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{658F9851-4D1C-4D57-904E-F9D7DA7F24FC}.Debug|x64.ActiveCfg = Debug|x64
-		{658F9851-4D1C-4D57-904E-F9D7DA7F24FC}.Debug|x64.Build.0 = Debug|x64
 		{658F9851-4D1C-4D57-904E-F9D7DA7F24FC}.Debug|x86.ActiveCfg = Debug|x86
 		{658F9851-4D1C-4D57-904E-F9D7DA7F24FC}.Debug|x86.Build.0 = Debug|x86
-		{658F9851-4D1C-4D57-904E-F9D7DA7F24FC}.Release|x64.ActiveCfg = Release|x64
-		{658F9851-4D1C-4D57-904E-F9D7DA7F24FC}.Release|x64.Build.0 = Release|x64
 		{658F9851-4D1C-4D57-904E-F9D7DA7F24FC}.Release|x86.ActiveCfg = Release|x86
 		{658F9851-4D1C-4D57-904E-F9D7DA7F24FC}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
