+++ TownSuite.TwainScanner/Program.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using System.Windows.Forms;
+using TownSuite.TwainScanner.Backends;
 
 namespace TownSuite.TwainScanner
 {
@@ -18,6 +19,7 @@ namespace TownSuite.TwainScanner
             string ocrApiUrl = "";
             string ocrBearerToken = "";
             string workingDir = Environment.GetEnvironmentVariable("TMP");
+            string backend = "originaltwain";
 
             List<string> scanlist = new List<string>();
             for (int i = 0; i <= Environment.GetCommandLineArgs().Length - 1; i++)
@@ -43,7 +45,7 @@ namespace TownSuite.TwainScanner
                 }
                 if (Environment.GetCommandLineArgs()[i] == "-ocrapiurl")
                 {
-                    ocrApiUrl = Environment.GetCommandLineArgs()[i+1];
+                    ocrApiUrl = Environment.GetCommandLineArgs()[i + 1];
                 }
                 if (Environment.GetCommandLineArgs()[i] == "-ocrbearertoken")
                 {
@@ -53,14 +55,21 @@ namespace TownSuite.TwainScanner
                 {
                     workingDir = Environment.GetCommandLineArgs()[i + 1];
                 }
+                if (Environment.GetCommandLineArgs()[i] == "-backend")
+                {
+                    backend = Environment.GetCommandLineArgs()[i + 1];
+                }
 
             }
+
+            string dirText = Path.Combine(workingDir, "TownSuiteScanner");
 
             if (ScanList == false)
             {
                 Application.EnableVisualStyles();
                 Application.SetCompatibleTextRenderingDefault(false);
-                Application.Run(new MainFrame(settings, new Ocr(enableOcr, ocrApiUrl, ocrBearerToken), workingDir));
+                Application.Run(new MainFrame(settings, new Ocr(enableOcr, ocrApiUrl, ocrBearerToken),
+                    dirText, backend));
             }
             else
             {
@@ -71,9 +80,6 @@ namespace TownSuite.TwainScanner
                 Console.WriteLine("ScanListEnd");
                 Console.Out.Flush();
             }
-
-
-
         }
     }
 }
